perf(user): add lang to user and rename classroom to schoolclass
