--- BACKEND/src/users/entities/class.entity.ts
@@ -1,13 +1,0 @@
-import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
-
-@Entity()
-export class Classroom {
-  @PrimaryGeneratedColumn('increment')
-  id: number;
-
-  @Column('character varying')
-  name: string;
-
-  @Column('character varying')
-  address: string;
-}
