--- BACKEND/src/users/entities/roles.entity.ts
@@ -1,10 +1,0 @@
-import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
-
-@Entity()
-export class Role {
-  @PrimaryGeneratedColumn('increment')
-  id: number;
-
-  @Column('character varying')
-  name: string;
-}
